requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Deprecated==1.2.14
distro==1.9.0
Django==5.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
envparse==0.2.0
idna==3.7
netaddr==1.3.0
phonenumbers==8.13.43
pillow==10.4.0
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
wrapt==1.16.0
yookassa==3.3.0
